backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.25.2
playwright==1.40.0
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Data Analysis
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
textblob==0.17.1
jieba==0.42.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1
loguru==0.7.2